backend/requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
sqlalchemy>=2.0,<3.0
aiosqlite>=0.20,<1.0
pydantic>=2.6,<3.0
pydantic-settings>=2.2,<3.0
jinja2>=3.1,<4.0
python-multipart>=0.0.9,<1.0
apscheduler>=3.10,<4.0
tenacity>=8.2,<9.0
httpx>=0.27,<1.0
pillow>=10.3,<12.0


